requirements.txt
pandas
numpy
scikit-learn
fastapi
uvicorn
pydantic
joblib
sqlalchemy
pymysql
passlib
bcrypt
python-multipart
email-validator
PyJWT
cryptography
